Rename to spandsp3
This renames the library to libspandsp3. The main header is also renamed
to spandsp3.h. The folder of the secondary headers is renamed to
spandsp3 likewise. This way libs and headers can co-exist with the older
spandsp version.

Programs using spandsp need to be adapted slightly to use the new
version, but this is acceptable as they need some code changes anyway.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>