configure.ac: uncomment AC_CANONICAL_BUILD
This commit uncomments `AC_CANONICAL_BUILD`.  This is needed in
order to make cross-compilation work properly (at least in nixpkgs).

The `AC_CANONICAL_BUILD` line in `configure.ac` was commented out in
2009 as part of a gigantic megapatch imported from subversion.  It
is unclear why it was commented out.  It appears to make the `if`
conditional (the next stanza of code) into deadcode.