Propagate CPPFLAGS_FOR_BUILD for build tools
On building build tools, only CC_FOR_BUILD is used with
CPPFLAGS_FOR_BUILD never actually passed. This was probably an oversight
when CPPFLAGS_FOR_BUILD was introduced.

This is especially needed on cross compilation where the header are
placed on a standard location.

To address this case, and make building build tool more robust, actually
pass the CPPFLAGS_FOR_BUILD for each tool.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>